+++ c/py_column.cc
@@ -43,10 +43,11 @@ int unwrap(PyObject* object, Column** address) {
 
 
 PyObject* column_from_list(PyObject*, PyObject* args) {
-  PyObject* list;
+  PyObject* arg1;
   int stype = 0, ltype = 0;
-  if (!PyArg_ParseTuple(args, "O!|ii",
-                        &PyList_Type, &list, &stype, &ltype)) return nullptr;
+  if (!PyArg_ParseTuple(args, "O|ii", &arg1, &stype, &ltype))
+    return nullptr;
+  PyyList list = PyObj(arg1);
   Column* col = Column::from_pylist(list, stype, ltype);
   return from_column(col, nullptr, 0);
 }