+++ c/py_datatable.h
@@ -177,8 +177,11 @@ DECLARE_METHOD(
 
 DECLARE_FUNCTION(
   datatable_from_list,
-  "datatable_from_list(...)\n\n"
-  "Create a DataTable from ...",
+  "datatable_from_list(list, types)\n\n"
+  "Create a DataTable from a list of Python objects (each will be converted\n"
+  "into a column). Optional `types` list may be provided to force a\n"
+  "particular stype for each column. The lengths of `types` list must be the\n"
+  "same as the number of entries in the source list.",
   HOMEFLAG)
 
 DECLARE_FUNCTION(